requirements.txt
imageio==2.34.1
numpy==1.26.4
pillow==10.3.0
